user savestates, saving/loading, no effect saving
Because I don't save the effect list, this produces a bug from C++ Lix:
When we load a user savestate from the very future and have no effects
in the EffectManager recorded, then framestep back, replay arrows are
shown during the on-the-fly forward recalculation. These arrows should
not be shown during framestepping back.