GadgetTile : Platonic, TerrainTile : Platonic
game.phymap -> tile.phymap, has 2 uses

Two uses of this class:
a) game uses it to look up physics.
b) terrain/steel uses a small one, from which a) is generated.

But watchout: Water/fire/flingers/exits/hatches etc. don't use a physicsmap
to specify their trigger areas! The map in a) is composed from the tile
definitions, not from the physicsmap in the tile.