editor, 2 instead of 4 fg/bg buttons
Need to draw undo and redo buttons.
Fg/bg are on the wrong IDs in the image, so they're reordered
in file.language right now as a workaround. Result of the
workaround: Undo/redo are not where I'd like them, but all
buttons do what's written and drawn on them.