main browser: dirs/files up/down by hotkey
This may be bad design. gui.picker.Picker could do it.
But Picker doesn't know the recent file, or the currently
highlit file or directory.

Idea to improve: Picker could countUntil it finds a highlit
entry, then assume that's the only highlit entry, and go up/down
from there. Picker should not report executeDir. Should it
report executeFile on mere up/down action? Would be good for the
BrowserBase browsers, but not for the level editor file picker.