fix #27: replay browser doesn't open 2 files
The replay browser displays the replay's filename, without extension.
Before, it opened the replay and the linked-to level file, to display
a nice button caption with the level title and player name.
I found that too slow in a long directory. Let's see whether the
bare filename of replays is too ugly.