fix #122: Continued replay points correctly
Save a replay, then load the replay from disk and play on.
You can save this as a replay, and it will point to the
correct file.

Yet to test: Autoreplay, what level is written into the replay,
decide whether we should get checkmarks from the verifier.

Simplified the code: Tribes don't know about their masters.