fix #157: old replays turned ticks into ?!
Buggy behavior in 0.6.16:

1. Solve level, this generates replay A. Observe how the level gets a tick.
2. Update the level. Observe how the level gets a ?!.
3. Solve the level again. Observe how the level gets a tick.
4. Play back the replay A.

Expected: Tick remains. Observed instead: ?! on the level.

Now, the replay browser checks whether the level is identical to the
pointedTo level, and only then allows overwriting checkmarks.

When you run levels from the command line with a replay, the replay will
only save ticks when you run it explicitly with the --pointedTo switch.
That's correct.

When you run levels from the command line, and explicitly tell the
level separately from the replay to Lix (by running with 2 file args),
you will always save checkmarks because played level is the level in
the only level file in question (the included level isn't used, neither
is the pointed-to level).

Bug left behind: When we don't run with pointed-to and pass only one
file argument, we will never save replays, not even when the included
level is identical to the (not considered) pointed-to level.