option titles and descs merged in lang
In the translation files in ./data/transl/, the options in the options
menu have a name and a long description. Both of these are now in the
same translation string: First comes the short name, then a pipe `|',
then the 1st line of the long description, then `|', then the 2nd line.