fix #302: observers were listed with players
In the end-of-game line in the lobby console, observers were listed
instead of players when the observer had PlNr 0.

Replays don't remember anymore who was local player and who was remote
player. When loading a replay in the interactive game, the user name
from the global configuration determines what side to observe.