fix #55: remove end-of-game dialog
We had issue #55: The end-of-game dialog should allow you to progress to
the next level immediately.

I believe that an adequate implementation of that wish should offer two
levels: 1) The next unsolved level and 2) the next level regardless of
solved status. It seems clearer to exit to the level browser immediately
and let the user pick the next level themself. Thus, I've removed the
end-of-level button, and subsequent commits will reintroduce its
functionality into the level browser after a level has been played.