trophies: save as SDLang, load as both
Erroneous trophy SDLang files are backed up. Data loss is prevented like
so: The game, after it couldn't load the broken file on startup, will
write, on exiting, an empty file to the normal SDLang file. You can
then fix the backup manually and rename it to the normal SDLang file,
to get your trophies back.