game.score -> gui.score
Reason for the move: The lobby shall eventually display score bars after
a game. The lobby shall get scores from the game, but not the entire GUI
widget. The lobby shall create the widget itself.

The lobby doesn't display any GUI widget for scores yet; instead, the
lobby still prints the scores to the chat.