Fix #379: Avoid cloning tiles outside map
Old behavior:
Usually, the editor moves freshly-cloned tiles down-right by (16, 16).
Sometimes, this would move tiles out of the map.

New behavior:
If a cloned tile would move out of the map, or even very close to the
edge (= undercutting a safety margin of 16 pixels), all cloned tiles
are moved up-left by (16, 16) instead. If neither offset keeps all
tiles on the map, we move down-right by (16, 16) and accept that
some are outside. In torus directions, tiles are always on the map
within the safety margin.