ScoreGraph: Removed black box on nuking players
This black box obscured the score in a ScoreGraph.
This box was a hack when we didn't have the larger ScoreBoard.

This black box was never present on ScoreBoard (what appears when you
mouse over the ScoreBar in the panel). The ScoreBoard shows nuke icons
instead. If you want to see who nuked, mouse over the ScoreGraph to see
the ScoreBoard with its nuke icons next to nuking players.