Multiplayer exiting dialog: Esc resumes
During networked play, it was easy to double-tap Esc, which quits
to the lobby. Now, this Esc dialog is more like the Editor's lose-data
dialog: The second press of Esc dismisses the dialog and returns to
play/editor. Thus, double-tapping Esc is no problem.

To exit the to lobby, you must first open the dialog by pressing Esc,
then press the hotkey for data loss (mapped by default to G and Delete)
that is harder to hit by accident.