Fix #401: Tweaker mispainted when plies swapped
This also removes a hackish caching variable.

It's still not perfect design because we must guess whether our child x
will change when we call a method on x. Thus, the current bad design is:

	if (x.y != y) {
	    x.y = y;
	    redraw ourselves;
	}

Still not perfect, but better, would be:

	x.y = y;
	...
	if any x has to be redrawn, redraw ourselves

But a direct implementation of this would require a public member
Element.drawRequired, which wouldn't be used for anything else.