Fix #385: Tweaker shows plies around NowLine
The original bug #385 was: Tweaker won't show more than 18 plies,
therefore it should get a scrollbar.

The solution is: Instead of a scrollbar, we show only the plies around
the NowLine. To scroll, move forward or backward in time, so that
different plies will now lie closest before/after the NowLine.