Comply with -preview=in
"void foo(in X)" means void "foo(scope const(X))".
Don't have "in ref", that is now a compiler error. "in" transforms to
reference or pointer under the hood, anyway, I believe, and should be
doing the right thing in the most performant way.

Don't take "in X" when the X will be stored; take a "const(X)" then.