Fix #417: Editor shows all teams per gadget
Lix's rule for multiplayer: When there are fewer exits/hatches than
players, the players will be associated round-robin 1:N with the
exits/hatches. But the editor printed only the first association.

Now, the editor prints all associations that will happen in play
not merely the first per hatch/exit.

Example: 4 players A, B, C, D and 2 exits. The editor printed on
the exits: A, B. The editor should instead print on the exits the
outcome of distributing the 4 players round-robin to the exits:
AC, BD.

It's still awkward when the gadget count and the player count don't
cleanly divide each other. Now, the editor will print AE, B, C, D
on 4 gadgets for 5 players, even though in play, all gadgets but
one would be removed to guarantee a clean division. The editor
doesn't warn about non-clean division; maybe the editor should.