Fix DMD build in release mode
The problem was: DMD failed with exit code -4 and no error message.
This happened in DMD 2.098.1. It happened only in the release build:

    dub build -b release --compiler=dmd --arch=x86_64

The debugging build was always fine.

DMD didn't like these particular comparisons of a string (that
had no [] at the end of the symbol) with a sliced range.
Both comparisons were in a final override of a bool-returning method.

Adding [] at end of the string symbol would have fixed the build, too.
Example: You could fix this line in file.filename.base by adding []:

    parent.rootless == rootless[0 .. parent.rootless.length];
    parent.rootless[] == rootless[0 .. parent.rootless.length];

Here is a reduction of that DMD bug; save it as a file foo.d, then
build it with this command: dmd -release -O foo.d

    interface I {
    immutable:
        string a() nothrow pure;
        string b() nothrow pure;

        final bool func(immutable(I) other)
        {
            return other.a && b[0 .. other.b.length] == other.b;
        }
    }