Fix lagging client in a 10-hour netgame
This fixes an esoteric bug that never surfaced in normal usage.
Single multiplayer games typically run for 30 seconds to 5 minutes,
not for 10 hours, thus typical games have never run into this bug.

The bug was: When the same networking game (not netplay session or
server run, but really individual game; what happens after you click
"ready" in the lobby) ran for over 10 hours, the client would stop
updating at normal speed and instead run at 66 % of that speed.

The reason was: int is a signed 32-bit integer, with max (2^31 - 1).
In the client, I kept milliseconds in an int. The computation
    millis * ticksPerSecond / 1000
will overflow when millis is greater than ((2^31 - 1) / 60), which
is 35791394 msecs = 35791 seconds = 9.94 hours. The Lix client later
interpreted the computation as the server's recommended time since start
of game, which, after overflow, was negative. Therefore, the Lix client
would continue to update at the slowest speed, 6 phyus per second
instead of the normal 4 phyus per second, to come closer to that target.

The fix is: Compute in D's standard library type core.time.Duration,
internally kept as hectonanoseconds stored in a long (64-bit integer).
This will work until a single game runs for 2^63 / 10^7 seconds
= 922337203685 seconds = 29247 years.