Make Lix compile with dmd 2.100.0, enum in ternary
Lix failed to compile in dmd 2.100.0, which was released this week.
Lix compiled in all earlier versions of dmd.

Before dmd 2.100.0, such code was fine:

    enum E { a, b }
    E e = true ? a : b;

(This particular code still works, it's not a reduction of Lix's fail to
compile. Lix had code similar, not equal to, this. I haven't reduced Lix
yet.)

Now, the compiler first interprets b as int 1, and then is unhappy that
the ternary expression is of type int, and doesn't want that to assign
to e.