Fix deprecations for dmd 2.100.0
Both of these functions masquerade a class union (a field long enough
to hold any one child class of a given base class) as the base class to
the outside.

The GC isn't responsible for those classes that I embed in the union.

I've annotated with return the functions that escape references.