Rename my older map to Fry-Day
This is to avoid a naming collision when talking about maps. Fry-Day
used to be called Rescue Rangers, but this name (Rescue Rangers)
appears in a far more popular map by Dominator101,
Rescue Rangers Double Team,
which also involves building to save a top hatch, but is otherwise
completely different and has nothing to do with Fry-Day.