Fix #387: Update skill panel on tweak
Now, after shifting a skill assignment forward or backward by a single
frame in the replay tweaker, the skill panel updates all skill counts
properly. This matters when some skills become unassignable in the
now-edited replay.

Cosmetical problem: When you shift assignments forward frame-by-frame,
it's possible that the game plays the assignment sound on every shift.
But that is unrelated to #387. The sound issue has always been in the
tweaker. The sound issue should become a new bug report.