Tribe implements handicap, Score shows handicap
Rational arithmetic for Score.lixSaved:
Format fractions with subscripts, don't print them with %.1f which would
produce the same figures, but too large a decimal digit. We always round
down in the decimal place.

Show raw lix alive number, but handicapped bar:
The number of lix alive will be printed as raw when it's printed as a
number, e.g., next to a lix icon in the status bar, or next to a lix
icon in the annotated bars on the big ingame score board.

The bars themselves in the big ingame score board are drawn with the
handicapped potential, not with the raw number of lix alive. This is
better because the length of the potential bar will stay constant when
you save your own lix.

We still allow replays without players. As in Lix 0.9, the Nurse base
class will add a player with the garden style to guarantee that the
GameState will always divide/modulo by >= 1.