Go back to lobby after 1-tribe multiplayer
It's possible to start a multiplayer game with only 1 team. The code
considers this singleplayer, but also considers it to be View.battle,
leading to odd results. In particular, before this fix here, the game
wouldn't terminate when that 1 tribe didn't save enough lix to meet the
level's save requirement. Instead, the console would print the save
count every frame. You had to press Esc to go back to the lobby.

Now, in this odd mode, we always go back to the lobby, even when we
didn't meet the save requirement that is nowhere shown in game.

Long-term idea: Implement full-fledged multiplayer puzzle solving
with framestepping and all other bells and whistles.