Abbreviate replay filename at beginning, not end
In the replay browser's nameplate, we abbreviate the filename at the
beginning: "Text too long" -> "...o long". This is because we abbreviate
the filename at the end everywhere else, and without this commit, there
would be no way to see the filename's tail anywhere in Lix.

The file picker in the replay browser's left-hand side still abbreviates
at the end of the filename: "Text too long" -> "Text too l."