Larger undraw area for Label "Jt"
Replay filenames end in .txt, ending in "t", and will show in the
browser with that ending because we abbreviate them from the front,
not the back.

The letter "J" will stick out of the left boundary, the letter "t" will
stick out of the right boundary of Label. I've increased the leftward
undrawing area of Label in the past. Now, I've increased the rightward
area, too.