Fix #386: Scroll to right edge with tweaker
The bug was: Open the replay tweaker. You were not able to scroll all
the way to the right side of the map. You had to zoom out to see the
entire map, or zoom in deeply enough to make the tweaker obscure less
of the level. This was because the map only anticipated the normal
panel on the screen, not the tweaker. The map expected to be able to
draw all the way to the right edge of the screen.

Now, even while the tweaker is open, you can scroll such that the
right end of the map is visible to the left of the tweaker.

Also, we fix how the tweaker was appearing for 1 frame on level
start. This bug has been in Lix for 4 years, but we never realized it
because the old tweaker didn't re-focus cameras when showing/hiding.