Restart previous level with replay
The end-of-singleplayer screen offers several levels: The previously
played level, and zero, one, or two next levels. If you choose the
previously played level, we'll start that level with the replay that
we just received from that level. You can watch your solution again, or
keep working on it.

To play the previous level without replay, you have a choice:
1. Start the previous level from the end-of-singleplayer screen,
    then cancel the replay immediately within the first second of
    play, before any lix spawn.
2. Or exit from the end-of-singleplayer screen to the singleplayer
    browser, then start the same level from there.

Next levels, we'll always start them without replay.

On instantiating a Nurse in the game, the Nurse refuses early
singleplayer nukes that the replay might hold. This responsibility
really belongs elsewhere: The replay shouldn't accept early singleplayer
nukes in the first place, or the game shouldn't export such early nukes
when somebody calls Game.replay() or Nurse.constReplay().