Fix #428: Update next-level cache after solving
The bug was:
    1. Have three unsolved singleplayer levels: A, B, and C,
        in this order, in direct succession.
    2. From the singleplayer browser, play B.
    3. Solve B.
    4. Go back to the singleplayer level browser, play A.
    5. Solve A, or exit A without solving.

Expected: End-of-level screen offers B as next level
and C as next unsolved level.

Observed: End-of-level screen offers only B.

Now, the end-of-level screen offers B as next level and C as next
unsolved, as expected.

The following internal design problem is still as before; it has no
impact on the user: {
    The cache for checkmarks and the cache to find the next unsolved
    level are two different caches. The checkmarks are properly
    updated after solving B. But the next-level cache is rebuilt only
    because I call it explicitly after updating the checkmarks.
    There is no universal command that ensures that both databases
    are updated in sync.
}

Maybe it's worth it to merge the two caches in the long term.