Replay insert mode as a user option, default off
By default, this is off. When it's off, assignments during replay
cut the replay, then add the assignment, as they have always done.

If checked, assignments cut only the future of the same lix: All
assignments to that same lix, and nukes. The replay will keep
assignments to different lix. You must check this in the options and
then it's on permanently during play. You can't disable it during play.
Click air to cancel the global future as usual.

I'm sure we can improve this mode later, add better options, and add
better UI during play.