Fix #453: Separate hardware resolution setting
Before:
Windowed mode and hardware fullscreen offered a pair of integers
in the options menu to tell Lix the resolution. The very same pair
of integers affected both screen modes.

Now:
Hardware fullscreen offers its own pair of integers, which specify
a possibly different resolution than the integers for windowed mode.

The default for windowed mode remains 1280x720.
The default for hardware fullscreen is 640x480.

Reason: I think that hardware fullscreen, in particular, must be
comptabile with as many screeens as possible. Thus, it's 640x480.

Hack around a bug in Allegro 5.2.8 on Linux: Hardware resolution
applies also to software fullscreen, to get a right-sized window.