Don't play music during name entry
Play the music only in the main menu. This is to not shock new players
during name entry if the music is too loud or unexpected.

It's easier to abort quickly from the main menu: You don't have to guess
that Esc quits, you can click a button to quit.