Allow queuing sounds with custom dBFS
Before, loud was dbToGain(x), and quiet was dbToGain(x) * 0.2f.

Now, loud is still dbToGain(x), but quiet is dbToGain(x - 12).
This is minimally more quiet; the true value would have been to subtract
(5 * log(5)/log(2)), which rounds to 11.60964, instead of 12. To see
that it's about the same, you could add a test:

    unittest {
        for (int dBFS = -50; dBFS < 30; dBFS += 5) {
            assert (dbToGain(dBFS - 12) > 0.15 * dbToGain(dBFS));
            assert (dbToGain(dBFS - 12) < 0.25 * dbToGain(dBFS));
        }
    }