Fix #448: Stunners don't repeat "Ouch" in flinger
The situation: Tumblers are stuck in a permanent flinger, but also
become a tumbler in every frame because of ground collision. They
oscillate between tumbler and stunner.

The bug was: They said "Ouch" every frame. That was annoying.

Now: They don't say "Ouch" when they stun inside the trigger area of a
permanent flinger. They still say "Ouch" in the trigger area of a
triggered flinger.