Fix #362: Play all in music/, don't filter .mp3
Don't exclude .mp3 because MP3 files play well on Windows with
contemporary Allegro, e.g., Allegro 5.2.8.0, if I don't exclude.

Allegro needs MiniMP3 on Linux or Mac, which is not provided in,
e.g, the Arch Linux Allegro 5 package. Thus, on Linux or Mac, you don't
necessarily have MP3 support from the get-go. That's not a bug in Lix.
But we should still remember this because people will consider it a bug
in Lix; we should then point people to the missig dependencies.

Thus: This fixes #362, "Warn about MP3 in main menu (unsupported but
common format)", without adding a warning, because we'll consider MP3
supported. On Windows, where it probably matters most, Allegro 5
supports MP3 very well in practice.