Fix truncated numbers at 1920x1080
The bug was:

During play, you have the skillbar at the bottom, and the infobar above
the skillbar. In the infobar, there is a hatch icon and a 3-digit count
for number of lix in hatch. At 1920x1080 screen resolution, number of
lix in hatch abbreviated its third digit to a dot ".", e.g., 123 showed
as "12." instead of as "123".

Similarly, in the topology menu in the editor, for width/height of a level,
the hexadecimal value got truncated to e.g. 0x13. instead of 0x13A,
and the 0x in the color select got truncated to .

Fix:

Widen the text field for number of skills in hatch. It overlaps with
the total-lix icon, but that icon is long and vertical, with
transparency at both left/right sides. You can't see overlap.

Similarly, widen the relevant text fields in the topology menu.
Similarly, widen lix ID and tick count in the tweaker lines.