Clicks on facing-filtered lix don't cut replay
Hold the hotkey to filter for left-facing lix. Hover over a right-facing
lix. Click the left mouse button. This will not assign (correct). But
before this commit, this cut the replay (surprising). Now, you must
click air (best) or a dying lix (strange but still allowed) to cut.

This obviates the need for the option whether air clicks should cut the
replay, because we'll have fewer misclicks now. Thus, this commit also
removes that option.