Fix #479: (Un-)Grouping preserves torus invariant
This fixes a crash from asserting consistent history in debug mode.

The torus invariant is: When a dimension is wrapped (the x-dimension,
the y-dimension, or both x and y for a torus map), all tiles on that
map must have coordinates in [0, L[ for that dimension, where L is
the map size in that dimension.

E.g., Hopscotch has size 256x160 and wraps only the y-dimension.
On Hopscotch, each tile must have a y-coordinate of 0, 1, 2, ...,
157, 158, or 159. It's not allowed to have y = -16, 160, or 500.