Fix #487: Preserve unknown user options
When we don't know an option during option loading, we keep the SDLang
tag as-is, and export it during option saving at the end of the file.

Ignore type-mismatching option lines. E.g., mouse speed should never
become 0, but on type mismatch, the old code defaulted to 0. Now, if
mouse speed is 10 and a line wants to set it to a string, it stays 10.