Fix #439: Show remaining spawn delay in panel
In a multiplayer match or a multiplayer replay, when you've selected a
spawn delay handicap, the in-game panel will show the remaining time
in seconds until your first spawn.

Without a spawn delay handicap, your 4 seconds until first spawn
are not shown in the panel. They're easy to anticipate.

Likewise, in singleplayer, the 1 second is not shown.