Explain gadgets on mouseover
When the mouse is on the land, we explain the hovered gadget in the
panel. The description can get long. We'll have to think about the
typesetting in the panel's infobar.

Move the Gadget hierarchy into physics/.

Gadget doesn't rely on Graphic anymore. This was the last type to rely
on Graphic; the mouse cursor and the level preview had already dropped
their dependency on Graphic. Remove class Graphic altogether.

Refactor foreachConstGadget.

Refactor save requirement into Tribe ruleset. Tribes know if they win.

Write tooltips in 4 languages.