Use A5's al_convert_mask_to_alpha() again
See #431: After several years of investigation, the pink-to-alpha
recoloring that was broken on macOS was fixed in Allegro 5 itself.
The fix was to disable Allegro 5's native image loaders on Mac.
The fix will be available in Allegro 5.2.11 and newer.

An old suspicion that turned out incorrect was that the Allegro 5
function al_convert_mask_to_alpha() had a bug. I wrote a custom
replacement for it in 2023, to debug #431. That didn't help nor hurt.

It's now time to remove my custom replacement. This removal is
reasonable on Mac and elsewhere, and with any Allgero 5.2.x.